Replace anchore/inline-scan with anchore/anchore-engine
The anchore/inline-scan is the anchore/anchore-engine plus Postgres and a more than 5GB vulnerability database which is not used at all, as the evaluation is done in the backend.

This PR replaces anchore/inline-scan with anchore/anchore-engine and adapts the container creating command to run exactly the same analysis (metadata extraction) using the lighter container.

The RUN_CMD variable has been removed, as it is unused.